# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kantin Rumah Kayu API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/v1/tenants": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "operationId": "211e3c0b41e29db47adb477222c4d8f9",
                "parameters": [
                    {
                        "name": "open_now",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Tenant"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tenants/{tenant}/menus": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "operationId": "08370a141480a0dec4880e3138f12093",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menus by tenant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Menu"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tenant/menus": {
            "get": {
                "tags": [
                    "Tenant Menus"
                ],
                "operationId": "c43a7279073bf2ecf69d9c4b919294b9",
                "responses": {
                    "200": {
                        "description": "Tenant menus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Menu"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tenant Menus"
                ],
                "operationId": "dc86b0efc9a932c6e3bfd7d4839c8220",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "tenant_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 15000
                                    },
                                    "category": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "photo_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Menu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/v1/tenant/menus/{menu}": {
            "delete": {
                "tags": [
                    "Tenant Menus"
                ],
                "operationId": "b47fdfb5e97d603725d36512e28b8283",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu deleted"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tenant Menus"
                ],
                "operationId": "25f4ddf57481998eb7ea584664a74374",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    },
                                    "category": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "photo_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "stock": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Menu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/v1/tenant/menus/{menu}/stock": {
            "patch": {
                "tags": [
                    "Tenant Menus"
                ],
                "operationId": "ef57684dc1ed45dbab2f1bfb16e7c699",
                "parameters": [
                    {
                        "name": "menu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "stock"
                                ],
                                "properties": {
                                    "stock": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Menu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tenant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Warung Nusantara"
                    },
                    "opens_at": {
                        "type": "string",
                        "example": "08:00:00"
                    },
                    "closes_at": {
                        "type": "string",
                        "example": "21:00:00"
                    }
                },
                "type": "object"
            },
            "Menu": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "tenant_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "example": "Nasi Goreng"
                    },
                    "price": {
                        "type": "integer",
                        "example": 25000
                    },
                    "category": {
                        "type": "string",
                        "example": "main"
                    },
                    "photo_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}